Changed the type of the variable t_priority from int to short.

+++ src/bask_task.h
@@ -9,7 +9,7 @@ struct bask_task
 {
 	unsigned int n; /* Only positive values. */
 	unsigned int t_id; /* Only positive values. */
-	int t_priority;
+	short t_priority; /* 2 Bytes should be more than enough for this. */
 	short t_flags;
 	char t_added[T_S_ADDED];
 	char t_finished[T_S_FINISHED];
@@ -21,11 +21,11 @@ struct bask_task
 int task_check_input (bask_core* tcore, char*, char*, char*, char*, int);
 
 void task_free_ll (struct bask_task**);
-int task_insert (struct bask_task**, unsigned int, unsigned int, int, int, int, char*, char*, char*, char*);
+int task_insert (struct bask_task**, unsigned int, unsigned int, int, short, int, char*, char*, char*, char*);
 int task_remove (struct bask_task**, unsigned int);
-int task_modificate (struct bask_task**, unsigned int, int, int, int, char*, char*, char*, char*);
+int task_modificate (struct bask_task**, unsigned int, int, int, short, char*, char*, char*, char*);
 
-int task_create (bask_core*, struct bask_task**, int, char*, char*);
+int task_create (bask_core*, struct bask_task**, short, char*, char*);
 void task_deactivate (struct bask_task**, unsigned int);
 int task_finish (struct bask_task**, unsigned int);
